+++ include/x308/App.hpp
@@ -1,11 +1,37 @@
 #pragma once
 
+#include <iosfwd>
+#include <memory>
+
 namespace x308 {
 
-class App {
+struct AppContext;
+
+enum class ApplicationState { created, initialized, running, stopping, stopped };
+
+class Application {
 public:
-    [[nodiscard]] int run(int argc, const char* const* argv) const;
+    Application();
+    Application(std::istream& input, std::ostream& output, std::ostream& error);
+    ~Application();
+
+    Application(const Application&) = delete;
+    Application& operator=(const Application&) = delete;
+
+    [[nodiscard]] int run(int argc, const char* const* argv);
+    [[nodiscard]] ApplicationState state() const noexcept;
+
+private:
+    void shutdown() noexcept;
+    [[nodiscard]] int finish(int exitCode) noexcept;
+
+    std::istream& input_;
+    std::ostream& output_;
+    std::ostream& error_;
+    ApplicationState state_{ApplicationState::created};
+    std::unique_ptr<AppContext> context_;
 };
 
-}  // namespace x308
+using App = Application;
 
+}  // namespace x308