+++ src/main.cpp
@@ -1,6 +1,5 @@
 #include "x308/App.hpp"
 
 int main(const int argc, const char* const* argv) {
-    return x308::App{}.run(argc, argv);
+    return x308::Application{}.run(argc, argv);
 }
-